Make building for 64bit explicit
I was building treesitter with the default VS tooling with a 32bit version of the compiler. By adding this flag you will get build errors when compiling with the 32bit version because it can not build the 64bit assemblies needed for .net core. This will make it clear for the person that builds the tree-sitter dll that the wrong version of the compiler is used